fix(blocky): set resources back
Thought blocky was going OOM but it was the CPU limit. Blocky
does not need a lot of CPU when running, but when starting up,
it needs to parse and process the denylists and that takes some
time. If you don't give it enough CPU, startup takes forever and
the pod can timeout.