chore(immich): change NFS share
I started with a Longhorn RWX mount, which uses NFS under the hood.
The main reason advantange of this at the time was that it gave me
backups to the NAS.

But as now the NAS has external backups, I'm okay with a NFS share
directly on the NAS, so Immich does not use the limited local storage
in Longhorn.