feat(home-assistant): replace ingress with route
This took quite a bit of work as I had to figure out how to use a
shared Middleware across namespaces. Turns out the Gateway API just
doesn't really let you, and imo a ReferenceGrant would make sense
here but this doesn't seem supported.

The hack is to put a middleware in each namespace. Luckily if this
is a chain, it can still reference other middleware in other namespaces.
But this does require the full <namespace>-<name>@kubernetescrd notation.